requirements.txt
streamlit
torch
accelerate
transformers
docling-core
huggingface_hub
Pillow
python-dotenv
numpy<2
PyMuPDF